Finish Distribution to Other servers

+++ storage/src/service/command_handlers.py
@@ -10,7 +10,7 @@ import jwt
 from asyncpg.exceptions import PostgresError, UniqueViolationError
 from src.adapters.s3 import AbstractS3Storage
 from src.core import exceptions
-from src.core.config import settings
+from src.core.config import settings, tz_now
 from src.domain import command_results, commands, events, models
 from src.service.uow import AbstractUnitOfWork
 from src.tools.hasher import PBKDF2PasswordHasher
@@ -24,9 +24,11 @@ async def get_many_cdn_servers(
     cmd: commands.GetManyCdnServers,
     uow: AbstractUnitOfWork,
 ):
+    logger.info('Getting CDN servers')
     async with uow:
         objs = await uow.cdn_servers.get_all()
 
+    logger.info(f'Get {len(objs)} CDN servers')
     return command_results.PositiveCommandResult([x.dict() for x in objs])
 
 
@@ -34,13 +36,14 @@ async def create_cdn_server(
     cmd: commands.CreateCdnServer,
     uow: AbstractUnitOfWork,
 ):
+    logger.info('Creating CDN server')
     async with uow:
         obj = await uow.cdn_servers.get_by_name(cmd.name)
 
         if obj:
             raise exceptions.CdnServerAlreadyExists
 
-        dt = datetime.now()
+        dt = tz_now()
         obj = models.CdnServer(
             **{**cmd.dict(), **{"id": uuid4(), "created": dt, "updated": dt}}
         )
@@ -48,12 +51,14 @@ async def create_cdn_server(
         await uow.cdn_servers.add(obj)
         await uow.commit()
 
+    logger.info(f'CDN server {obj.dict()} has been created')
     return command_results.PositiveCommandResult(obj.dict())
 
 async def update_cdn_server(
     cmd: commands.UpdateCdnServer,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f'Updating CDN server with id {cmd.id}')
     async with uow:
         server = await uow.cdn_servers.get_by_id(cmd.id)
 
@@ -61,18 +66,21 @@ async def update_cdn_server(
             raise exceptions.CdnServerDoesNotExist
 
         obj = models.CdnServer(
-            **{**server.dict(), **cmd.dict(), **{"updated": datetime.now()}}
+            **{**server.dict(), **cmd.dict(), **{"updated": tz_now()}}
         )
 
         await uow.cdn_servers.update(obj)
         await uow.commit()
 
+    logger.info(f'CDN server {obj.dict()} has been updated')
+    uow.push_message(events.CdnServerUpdated(id=server.id))
     return command_results.PositiveCommandResult(obj.dict())
 
 async def delete_cdn_server(
     cmd: commands.DeleteCdnServer,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f'Deleting CDN server with id {cmd.id}')
     async with uow:
         obj = await uow.cdn_servers.get_by_id(cmd.id)
 
@@ -82,16 +90,48 @@ async def delete_cdn_server(
         await uow.cdn_servers.delete(cmd.id)
         await uow.commit()
 
+    logger.info(f'CDN server {cmd.id} has been deleted')
     return command_results.PositiveCommandResult({})
 
+async def enrich_cdn_server_by_files(
+    cmd: commands.EnrichCdnServerByFiles,
+    uow: AbstractUnitOfWork,
+):
+    logger.info(f'Enriching CDN server with id {cmd.id} by files')
+    async with uow:
+        new_server = await uow.cdn_servers.get_by_id(cmd.id)
+        if new_server.is_on and new_server.is_active:
+            servers_in_zone = await uow.cdn_servers.get_switched_on(True, new_server.zone)
+            servers_ids_in_zone = [x.id for x in servers_in_zone]
+
+            files_on_servers = {}
+            for server_id in servers_ids_in_zone:
+                if server_id != new_server.id:
+                    files_on_servers[server_id] = await uow.files.get_non_deleted_on_servers([server_id])
+
+            files_on_new_server = await uow.files.get_non_deleted_on_servers([new_server.id])
+            files_ids_on_new_server = set([x.id for x in files_on_new_server])
+
+            for server_id in files_on_servers:
+                for file in files_on_servers[server_id]:
+                    if file.id not in files_ids_on_new_server:
+                        logger.info(f'Order file {file.id} to download from CDN server {server_id}')
+                        uow.push_message(commands.OrderFileToDownload(
+                            file_id=file.id,
+                            server_id=server_id
+                        ))
+
+    logger.info(f'CDN server with id {cmd.id} has been enriched by files')
 
 async def get_many_files(
     cmd: commands.GetManyFiles,
     uow: AbstractUnitOfWork,
 ):
+    logger.info('Getting files')
     async with uow:
         objs = await uow.files.get_non_deleted(cmd.user_id)
 
+    logger.info(f'Get {len(objs)} files')
     return command_results.PositiveCommandResult(objs)
 
 
@@ -99,6 +139,7 @@ async def delete_file(
     cmd: commands.DeleteFile,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f'Deleting file with id {cmd.id}')
     async with uow:
         obj = await uow.files.get_by_id(cmd.id)
 
@@ -117,6 +158,7 @@ async def delete_file(
         )
         await uow.commit()
 
+    logger.info(f'File {cmd.id} has been deleted')
     uow.push_message(events.FileDeleted(id=cmd.id))
     return command_results.PositiveCommandResult({})
 
@@ -125,6 +167,7 @@ async def rename_file(
     cmd: commands.RenameFile,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f'Renaming file with id {cmd.id}')
     async with uow:
         obj = await uow.files.get_by_id(cmd.id)
 
@@ -134,7 +177,6 @@ async def rename_file(
         old_name = obj.name
 
         obj.name = cmd.name
-        obj.updated = datetime.now()
 
         try:
             await uow.files.update(obj)
@@ -152,6 +194,7 @@ async def rename_file(
         )
         await uow.commit()
 
+    logger.info(f'File {cmd.id} has been renamed')
     return command_results.PositiveCommandResult(obj.dict())
 
 
@@ -159,6 +202,7 @@ async def get_file_servers(
     cmd: commands.GetFileServers,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f'Getting servers where file {cmd.id} is located')
     async with uow:
         objs = []
         ids_of_servers = await uow.files.get_ids_of_servers(cmd.id)
@@ -167,6 +211,7 @@ async def get_file_servers(
             servers = await uow.cdn_servers.get_all()
             objs = [x for x in servers if x.id in ids_of_servers]
 
+    logger.info(f'File {cmd.id} is located on {len(objs)} servers')
     return command_results.PositiveCommandResult(objs)
 
 
@@ -181,10 +226,10 @@ async def get_upload_link(
             f"Get upload link for file {cmd.name} on server {nearest_server.name}"
         )
 
-        obj = await uow.files.get_by_name_and_user_id(cmd.name, cmd.user_id)
+        obj = await uow.files.get_non_deleted_by_name_and_user_id(cmd.name, cmd.user_id)
         is_new = False if obj else True
 
-        dt = datetime.now()
+        dt = tz_now()
 
         if is_new:
             obj = models.File(
@@ -192,6 +237,7 @@ async def get_upload_link(
                 name=cmd.name,
                 size=cmd.size,
                 user_id=cmd.user_id,
+                has_deleted=False,
                 created=dt,
                 updated=dt,
             )
@@ -227,6 +273,7 @@ async def get_user_actions(
     cmd: commands.GetUserActions,
     uow: AbstractUnitOfWork,
 ):
+    logger.info('Getting user actions')
     async with uow:
         objs = await uow.history.get(cmd.user_id, cmd.limit, cmd.offset)
         count = await uow.history.count(cmd.user_id)
@@ -297,6 +344,7 @@ async def handle_s3_event(
     cmd: commands.HandleS3Event,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Handling S3 event with key {cmd.routing_key}, body {cmd.body}")
     async with uow:
         try:
             storage_name = cmd.routing_key.split(".", 2)[1]
@@ -328,6 +376,8 @@ async def handle_service_event(
     cmd: commands.HandleServiceEvent,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Handling service event with key {cmd.routing_key}, body {cmd.body}")
+
     key = cmd.routing_key.split(".", 1)[1]
 
     events_mapping = {
@@ -348,12 +398,29 @@ async def mark_file_as_stored(
     cmd: commands.MarkFileAsStored,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Marking file with id {cmd.file_id} as stored")
     async with uow:
         await uow.files.add_server_to_file(cmd.file_id, cmd.server_id)
-        if await uow.files.is_copied(cmd.file_id):
-            uow.push_message(events.FileDistributed(id=cmd.file_id))
 
+        server = await uow.cdn_servers.get_by_id(cmd.server_id)
+        ids_of_servers_where_file_located = await uow.files.get_ids_of_servers(cmd.file_id)
+        ids_of_servers_in_zone = [x.id for x in await uow.cdn_servers.get_switched_on(True, server.zone)]
+        
         await uow.commit()
+
+    # print('***** MARK FILE AS STORED ******')
+    # print(ids_of_servers_in_zone)
+    # print(ids_of_servers_where_file_located)
+    # print(set(ids_of_servers_in_zone) - set(ids_of_servers_where_file_located))
+    # print(len(set(ids_of_servers_in_zone) - set(ids_of_servers_where_file_located)))
+
+    if len(set(ids_of_servers_in_zone) - set(ids_of_servers_where_file_located)) == 0:
+        uow.push_message(events.FileDistributed(id=cmd.file_id))
+    else:
+        if len(ids_of_servers_where_file_located) == 1:
+            uow.push_message(commands.OrderFileToDownload(
+                file_id=cmd.file_id, server_id=cmd.server_id
+            ))
 
     return command_results.PositiveCommandResult({})
 
@@ -362,6 +429,7 @@ async def order_file_to_download(
     cmd: commands.OrderFileToDownload,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Ordering file to download")
     message = models.FileOrderedToDownloadBrokerMessage(message=cmd.dict())
     uow.push_message(commands.PublishMessage(message=message))
     return command_results.PositiveCommandResult({})
@@ -371,14 +439,37 @@ async def download_file_to_temp_storage(
     cmd: commands.DownloadFileToTempStorage,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Downloading file with id {cmd.file_id} from server with id {cmd.server_id}")
     async with uow:
         server = await uow.cdn_servers.get_by_id(cmd.server_id)
         storage = await uow.s3_pool.get(server.name, server.host, server.port)
         filename = str(cmd.file_id)
         path = os.path.join(settings.storage_path, filename)
         await storage.download_file(filename, path)
-        uow.push_message(events.FileDownloadedToTempStorage(id=cmd.file_id))
 
+    uow.push_message(events.FileDownloadedToTempStorage(id=cmd.file_id, zone=server.zone))
+    logger.info(f"File with id {cmd.file_id} has been downloaded to temp storage")
+    return command_results.PositiveCommandResult({})
+
+async def distribute_file_within_zone(
+    cmd: commands.DistributeFileWithinZone,
+    uow: AbstractUnitOfWork,
+):
+    logger.info(f"Distributing file with id {cmd.file_id} with own zone")
+    async with uow:
+        file_on_servers_ids = await uow.files.get_ids_of_servers(cmd.file_id)
+        servers_in_zone = await uow.cdn_servers.get_switched_on(True, cmd.zone)
+
+        for server in servers_in_zone:
+            if server.id not in file_on_servers_ids:
+                uow.push_message(
+                    commands.OrderFileToCopy(
+                        file_id=cmd.file_id,
+                        server_id=server.id,
+                    )
+                )
+
+    logger.info(f"File with id {cmd.file_id} has been distributed")
     return command_results.PositiveCommandResult({})
 
 
@@ -386,6 +477,7 @@ async def order_file_to_copy(
     cmd: commands.OrderFileToCopy,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Ordering file with id {cmd.file_id} to copy to server with id {cmd.server_id}")
     message = models.FileOrderedToCopyBrokerMessage(message=cmd.dict())
     uow.push_message(commands.PublishMessage(message=message))
     return command_results.PositiveCommandResult({})
@@ -395,6 +487,7 @@ async def copy_file(
     cmd: commands.CopyFile,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Coping file with id {cmd.file_id} to copy to server with id {cmd.server_id}")
     async with uow:
         server = await uow.cdn_servers.get_by_id(cmd.server_id)
         storage = await uow.s3_pool.get(server.name, server.host, server.port)
@@ -402,6 +495,7 @@ async def copy_file(
         path = os.path.join(settings.storage_path, filename)
         await storage.upload_file(filename, path)
 
+    logger.info(f"File with id {cmd.file_id} has been copied to server with id {cmd.server_id}")
     return command_results.PositiveCommandResult({})
 
 
@@ -409,12 +503,19 @@ async def remove_file_from_temp_storage(
     cmd: commands.RemoveFileFromTempStorage,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Removing file with id {cmd.file_id} from temp storage")
     filename = str(cmd.file_id)
     path = os.path.join(settings.storage_path, filename)
     try:
+        for root, dirs, files in os.walk(path):
+            print(root, dirs, files)
+
         os.remove(path)
+        logger.info(f"File with id {cmd.file_id} has been removed temp storage")
+
     except FileNotFoundError:
-        raise exceptions.FileNotFoundInTempStorage
+        logger.warning(f"File with id {cmd.file_id} was not found in temp storage")
+        pass
 
     return command_results.PositiveCommandResult({})
 
@@ -423,6 +524,7 @@ async def order_file_to_remove(
     cmd: commands.OrderFileToRemove,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Ordering file with id {cmd.file_id} to remove from server with id {cmd.server_id}")
     message = models.FileOrderedToRemoveBrokerMessage(message=cmd.dict())
     uow.push_message(commands.PublishMessage(message=message))
     return command_results.PositiveCommandResult({})
@@ -432,11 +534,15 @@ async def remove_file(
     cmd: commands.RemoveFile,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Removing file with id {cmd.file_id} from server with id {cmd.server_id}")
     async with uow:
         server = await uow.cdn_servers.get_by_id(cmd.server_id)
-        storage = await uow.s3_pool.get(server.name, server.host, server.port)
-        await storage.remove_file(str(cmd.file_id))
 
+        if server.is_on:
+            storage = await uow.s3_pool.get(server.name, server.host, server.port)
+            await storage.remove_file(str(cmd.file_id))
+
+    logger.info(f"File with id {cmd.file_id} has been removed from server with id {cmd.server_id}")
     return command_results.PositiveCommandResult({})
 
 
@@ -444,6 +550,7 @@ async def mark_file_as_removed(
     cmd: commands.MarkFileAsRemoved,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Marking file with id {cmd.file_id} as removed from server with id {cmd.server_id}")
     async with uow:
         await uow.files.remove_server_from_file(cmd.file_id, cmd.server_id)
         await uow.commit()
@@ -462,6 +569,7 @@ async def create_file_share_link(
     cmd: commands.CreateFileShareLink,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Creating share link for file with id {cmd.file_id}")
     async with uow:
         file = await uow.files.get_by_id(cmd.file_id)
         if not file or file.user_id != cmd.user_id:
@@ -493,6 +601,7 @@ async def create_file_share_link(
 
         await uow.commit()
 
+    logger.info(f"Share link for file with id {cmd.file_id} has been created")
     return command_results.PositiveCommandResult(
         make_share_link_result(obj, file)
     )
@@ -502,6 +611,7 @@ async def get_file_share_links(
     cmd: commands.GetFileShareLinks,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Gettings share links for file with id {cmd.file_id}")
     async with uow:
         file = await uow.files.get_by_id(cmd.file_id)
         if not file or file.user_id != cmd.user_id:
@@ -509,6 +619,7 @@ async def get_file_share_links(
 
         objs = await uow.file_share_links.get_many(cmd.file_id)
 
+    logger.info(f"Get {len(objs)} share links for file with id {cmd.file_id}")
     return command_results.PositiveCommandResult(
         [make_share_link_result(x, file) for x in objs]
     )
@@ -518,6 +629,7 @@ async def get_file_share_link(
     cmd: commands.GetFileShareLink,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Getting single share links for file with id {cmd.file_id}")
     async with uow:
         file = await uow.files.get_by_id(cmd.file_id)
         if not file:
@@ -525,10 +637,11 @@ async def get_file_share_link(
 
         link = await uow.file_share_links.get(cmd.file_id, cmd.link_id)
         if not link or (
-            link.expire_at is not None and link.expire_at <= datetime.now()
+            link.expire_at is not None and link.expire_at <= tz_now()
         ):
             raise exceptions.FileShareLinkDoesNotExist
 
+    logger.info(f"Info about share link {cmd.link_id} for file with id {cmd.file_id} has been returned")
     return command_results.PositiveCommandResult(
         make_share_link_result(link, file)
     )
@@ -538,6 +651,7 @@ async def delete_file_share_link(
     cmd: commands.DeleteFileShareLink,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Deleting share link {cmd.link_id} for file with id {cmd.file_id}")
     async with uow:
         file = await uow.files.get_by_id(cmd.file_id)
         if not file or file.user_id != cmd.user_id:
@@ -560,6 +674,7 @@ async def delete_file_share_link(
         )
         await uow.commit()
 
+    logger.info(f"Share link {cmd.link_id} for file with id {cmd.file_id} has been deleted")
     return command_results.PositiveCommandResult({})
 
 
@@ -567,6 +682,7 @@ async def validate_file_share_link(
     cmd: commands.ValidateFileShareLink,
     uow: AbstractUnitOfWork,
 ):
+    logger.info(f"Validating share link {cmd.link_id} for file with id {cmd.file_id}")
     async with uow:
         file = await uow.files.get_by_id(cmd.file_id)
         if not file:
@@ -575,7 +691,7 @@ async def validate_file_share_link(
         link = await uow.file_share_links.get(cmd.file_id, cmd.link_id)
 
         if not link or (
-            link.expire_at is not None and link.expire_at <= datetime.now()
+            link.expire_at is not None and link.expire_at <= tz_now()
         ):
             raise exceptions.FileShareLinkDoesNotExist
 
@@ -585,4 +701,5 @@ async def validate_file_share_link(
             ):
                 raise exceptions.AuthNoPermissionException
 
+    logger.info(f"Share link {cmd.link_id} for file with id {cmd.file_id} has been validated")
     return command_results.PositiveCommandResult({})
